pyproject.toml
[project]
name = "stock-advisor"
version = "0.1.0"
description = "Multi-agent stock advisory system with Bull/Bear/Judge debate engine"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # LLM clients
    "anthropic>=0.40.0",
    "openai>=1.55.0",  # for embeddings only
    "httpx>=0.27.0",   # for xAI (Grok) raw API calls
    # Data sources
    "polygon-api-client>=1.14.0",
    "yfinance>=0.2.50",
    # Database
    "supabase>=2.10.0",
    # Technical analysis
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "ta>=0.11.0",
    # Config & utilities
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",  # retries
    "structlog>=24.4.0",  # structured logging
    "lxml>=6.0.3",
    "apscheduler>=3.11.2",
    "pytz>=2026.1.post1",
    "tzdata>=2026.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
